+++ prometheus/ready.go
@@ -3,19 +3,81 @@ package prometheus
 import (
 	"context"
 	"net/http"
-	"net/url"
 	"path"
 	"time"
 
-	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/config"
 	"github.com/pkg/errors"
 )
 
-func WaitForReady(inCtx context.Context, logger log.Logger, promURL *url.URL) error {
-	u := *promURL
-	u.Path = path.Join(promURL.Path, "/-/ready")
+const (
+	scrapeIntervalName = config.PrometheusTargetIntervalLengthName
+)
+
+func completedFirstScrapes(inCtx context.Context, cfg config.PromReady) error {
+	u := *cfg.PromURL
+	u.Path = path.Join(u.Path, "/metrics")
+
+	ctx, cancel := context.WithTimeout(inCtx, config.DefaultHealthCheckTimeout)
+	defer cancel()
+
+	mon := NewMonitor(&u)
+	res, err := mon.Get(ctx)
+	if err != nil {
+		return err
+	}
+	summary := res.Summary(scrapeIntervalName)
+	foundLabelSets := summary.AllLabels()
+	if len(foundLabelSets) == 0 {
+		return errors.New("waiting for the first scrape(s) to complete")
+	}
+
+	// Prometheus doesn't report zero counts. We expect absent
+	// timeseries, not zero counts, but we test for Count() != 0 on
+	// the retrieved metrics for added safety below.
+
+	if len(cfg.ScrapeIntervals) == 0 {
+		// If no intervals are configured, wait for the first one.
+		//
+		// TODO: We can't be sure there are only one interval.
+		// After time passes, we can check again--if any new
+		// intervals are discovered, print a warning about
+		// configuring intervals via --prometheus.scrape-interval
+		// to ensure safe startup.
+		for _, ls := range foundLabelSets {
+			if summary.For(ls).Count() != 0 {
+				return nil
+			}
+		}
+
+		return nil
+	}
+
+	// Find all the known intervals.
+	foundWhich := map[string]bool{}
+	for _, ls := range foundLabelSets {
+		for _, l := range ls {
+			if l.Name == "interval" && summary.For(ls).Count() != 0 {
+				foundWhich[l.Value] = true
+				break
+			}
+		}
+	}
+
+	for _, si := range cfg.ScrapeIntervals {
+		ts := si.String()
+		if !foundWhich[ts] {
+			return errors.Errorf("waiting for scrape interval %s", ts)
+		}
+	}
+
+	return nil
+}
+
+func WaitForReady(inCtx context.Context, cfg config.PromReady) error {
+	u := *cfg.PromURL
+	u.Path = path.Join(u.Path, "/-/ready")
 
 	// warnSkipped prevents logging on the first failure, since we
 	// will try again and this lets us avoid the first sleep
@@ -40,18 +102,27 @@ func WaitForReady(inCtx context.Context, logger log.Logger, promURL *url.URL) er
 				defer resp.Body.Close()
 			}
 
-			if err == nil && resp.StatusCode/100 == 2 {
-				return true
+			respOK := err == nil && resp.StatusCode/100 == 2
+
+			if respOK {
+				// Great! We also need it to have completed
+				// a full round of scrapes.
+				err = completedFirstScrapes(inCtx, cfg)
+				if err == nil {
+					return true
+				}
 			}
 
 			if !warnSkipped {
 				warnSkipped = true
 				return false
 			}
-			if err != nil {
-				level.Warn(logger).Log("msg", "Prometheus readiness check", "err", err)
+			if respOK {
+				level.Warn(cfg.Logger).Log("msg", "Prometheus /metrics scrape", "err", err)
+			} else if err != nil {
+				level.Warn(cfg.Logger).Log("msg", "Prometheus readiness check", "err", err)
 			} else {
-				level.Warn(logger).Log("msg", "Prometheus is not ready", "status", resp.Status)
+				level.Warn(cfg.Logger).Log("msg", "Prometheus is not ready", "status", resp.Status)
 			}
 			return false
 		}()