+++ CHANGELOG.md
@@ -8,6 +8,16 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Unreleased
 
+### Added
+- Sidecar waits for the first scrapes to complete before entering its 
+  run state. (#134)
+- New setting `--prometheus.scrape-interval` supports configuring scrape
+  interval(s) to wait for at startup. (#134)
+  
+### Removed
+- The `--startup.delay` setting has been removed in favor of monitoring when 
+  Prometheus actually finishes its first scrapes. (#134)
+
 ## [0.17.0](https://github.com/lightstep/opentelemetry-prometheus-sidecar/releases/tag/v0.17.0) - 2021-02-23
 ### Added
 - Automatically set (the same) `service.instance.id` for Destination/Diagnostics